Stub AdvicePlans::ChangeOwnershipCommand
We are sharing sessions with the desktop website. The Advice Plans
engine stores command objects in the session. Rails attempts to fetch
the whole session graph causing a
`class AdvicePlans::ChangeOwnershipCommand not a struct` error.

This commit introduces a stub for that object that allows us to continue
to share the session.