Remove invalid/unused RequireJS paths
The path alias for the entry-point script for an engine is constructed
using using the value of the `data-engine` attribute from the engine's
layout with "Config" added on the end (see bottom of application.js).

So these two paths for car_cost_tool and pensions_calculator we never
being used and in fact referred to non-existent files.

See: https://trello.com/c/PlIsKwOm