Refactor corporate decorator and spec
This changes the `CorporateDecorator` to accommodate for the corporate
tool directory and corporate tool pages. Previously, we would have had
to create a new partial for every individual tool we made available for
syndication. This change just uses a single partial for these types of
tool pages and an explicit partial for the directory displaying all
tools. All other articles with dynamic partials will follow normal
behaviour of needing to be created on content and have a relevant
`slug.html.erb` partial in `frontend`