Remove contact-us static page and accoutrements
The goal is to allow contact-us to be handled by the CMS. The issue is
that this page is a lynch-pin for the machinery that serves out static
pages. Without it tests fail as there's nothing left to serve out, so
this change involves removing the static page machinery and the related
tests.