Fix cucumber scenario by turning off feature flag
Obs.: This test is relying on the feature flat to be disabled,
but on production is enabled. Since there are no disabled-feature
on cucumber I am disabling on the features.yml and creating
an user story to revisit this and do a proper test when the
feature is disabled or enabled.