Add cache to the development mode using an env var
So there are a lot of times which we are not developing anything that integrates
with external APIS and since the development environment does not have
cache the page load becames slow every single time. This commit fixes that
by adding the possibility to enable cache in development every time is
necessary.

Obs.: I added the .env-example the cache false as default because
the cache should be enabled intentionally by everyone that is running
the application.