Add site search gem and replacing google search
Uses site search with git master for now
Rewriten decorators specs using site search gem
The benefits of this gem is that we switch off from any service
and implementing new services for site search in a very easy way

There were some features tests that were deliberated deleted for two reasons.

First: the spelling suggestion was kind trick to implement on time constraint
that we had so I remove all the feature and I aligned with the product
managers about it.

Second: the browsing pagination features were totally Google based so
they kind need of a rewrite to be more flexible. I will do a separate PR for
this.