Update date of birth validations and Ensure it is a date object on save
This is related to the PII encryption through attr_encrypted gem
During the encryption of this field when some of the tests were failing
the validations weren't entirely correct and so an extra validation was added to
make sure it was a valid date.

After this I also added a method to ensure the object being saved into the DB was
a date object. Before encryption this was handled as the field in the schema was a date object
However now that the encrypted field is a long string it needs to be converted for the Marshal
to ensure a date object is returned when being decrypted.