Fix reset password token
After the encryption PR, the reset password token stopped working.

The reason is the overwrite of the find_first_by_auth_conditions
on the user model.

This method is used on Devise for authentication and also for
the reset password token. Since this overwritten was just searching by email
then it explains why was not searching by token as well.

Making the conditions more general makes both of scenarios to work.