Make sure partner tools is loaded by each environment
The test environments appeared to be serving the production version partner-tools
syndication javascript.

The partner site needs to add this code:

<html>
  <head>
  </head>
<body>
  <div>
     <noscript><p>You have javascript turned off. Please turn javascript on to see the tools in action.</p></noscript>
     <p>
       <a id="debt_advice_locator_f2f" class="mas-widget" target="_blank" href="https://moneyadviceservice.org.uk/en/tools/debt-advice-locator">Where to go to get free debt advice</a>
     </p>
  </div>
<script class='application_scripts' src="https://moneyadviceservice.org.uk/assets/syndication/tools.js"></script>
</body>
</html>

By controlling the src of the domain of the syndication tools on the tools.js
we need to hard code the environments that we need to point out and
returning production as default.