Add example value to ATTR_CRYPT_KEY environment variable
This env var value is used to encrypt certain attributes on the User
model. The encryption logic requires it to be 32 bytes long. Not setting
a value of that length causes errors when attempting to create a User.
The most immediate consequence of this is mass test failures.

Set a dummy value of the appropriate byte length, so anyone setting the
project up for the first time doesn't get swamped with these errors on
first run of the test suite.