Change the example ATTR_CRYPT_KEY to 32-byte string
Before Ruby 2.4, OpenSSL::Cipher would simply truncate key lengths that
were too long. In 2.4+ it raises an error. The attr_encrypted
functionality in the User model breaks as a result of this unless the
ATTR_CRYPT_KEY is set to the correct length.

Update the example with a working value.