Add relative_url_root config to fix Devise redirect bug
A recent upgrade to Devise 4.6.1 broke authentication redirects. When
accessing an authenticated endpoint, eg - '/users/edit', a user should
be redirected to the sign in page. Following the upgrade, the user sees
an error instead.

The failing code is in Devise::FailureApp#scope_url. This method has
been updated in 4.6.1 to correctly handle cases where Devise is embedded
in another engine. Although this isn't our use case, we execute the code
anyway - it's written in a way that suggests an attempt at backwards
compatibility.

The purpose of scope_url is to return a redirect URL that can be sent to
the user. This includes setting a value in opts[:script_name], which
determines the root for any subsequently used path helpers. The method
checks the host app's Rails config for a relative_url_root. If it
doesn't find one, it uses the root_path helper as a fallback (see line
157). This results in an error as the locale is missing.

There are a couple of issues here:

- For reasons I haven't been able to determine, the default_url_options
  specified in the host app and referenced by Devise::FailureApp aren't
  used when calling the root_path helper.
- Even if they were, the result of scope_url would be incorrect. As
  script_name is supposed to be the root which the path helpers build
  upon, passing in the locale when we invoke root_path would produce
  a redirect URL like '/en/en/users/sign_in'.

I see three ways of resolving this, along with their corresponding
downsides:

- Add a custom FailureApp to Frontend that inherits from
  Devise::FailureApp and overrides the scope_url method.

  Diverges from the Devise codebase and introduces potential
  complexity during future upgrades of the library.

- Set '/' as the root_path, changing the path helper for '/:locale' to
  localised_path (or similar).

  We'd need to use the appropriate path helper across the codebase
  depending on which root we're referring to.

- Add relative_url_root to the Rails config, set to a value of '/'.

  The presence of this config is somewhat unexpected, as it's typically
  only used for serving up the app from a subdirectoy (ie - all routes
  have a prefix of '/app1').

Of the three, I've opted for the addition of relative_url_root. It's the
least complex, it doesn't alter the existing path handling of the app,
and the explanatory comment clarifies its presence in the configuration.