Update password reset email
Alters the password reset email to include the correct links, directly
to MoneyHelper. These are hard-coded for now but should be configured
per-environment when the time comes.

I've updated the production configuration to include the `www` subdomain
as it appears we don't redirect canonically now.

It's worth noting the Welsh changes are here for compatibility though we
don't actually send the Welsh email under any circumstances.