Add a redirector route to get around the Safari iframe cookie issue
Because Safari doesn't allow cross-domain cookies to be set in iframes without first
settings a cookie on that domain (outside of an iframe) we need some sort of workaround.

This PR introduces a redirecter controller that looks like this:

/en/tools/redirecter/TOOL_NAME

So, in AEM we link to that path when we want to display a tool. The user clicks this link
and in turn get taken to partner-tools-moneyhelper.org.uk. This url sets a cookie, then redirects
back to the tool on AEM. This should mean that the cookie is now set on partner-tools.moneyhelper.org.uk
which means that iframes with said domain are now free to set cookies.