Resolve redirects after account actions
Ensure we never attempt to redirect back to the root path when an
account action is completed. The root path will always be a partner
tools URL that then issues further redirects back to MoneyHelper. In the
case of password resets, we need to redirect back into the embedded
environment as due to constraints when passing tokens via AEM pages, we
have to link directly to the password reset journey outside of AEM and
redirect back into it upon successful reset.