Check for cookies in syndicated/configured tools
This adds cookie detection in tools configured with `CookiedTool`
instances. For those tools that match a landing path, we redirect with a
cookie in the response and check it's set in the next request. If it is
we redirect back to the tool's landing path with the `checked` param set
to `true` to avoid a redirect loop. We also delete the initial cookie.

If the cookies cannot be set, we instead redirect to a page with a link
to the canonical (AEM hosted) tool that will open in a new browser
instance.

This is primarily to avoid users of syndicated tools from seeing
unexpected behaviour in situations where third-party cookies are
blocked. Primarily safari for now but other browser engines are due to
become stricter soon.