Only check landing pages on GET
We were indiscriminantly checking before any action, for those tools
that were configured for cookie checks. Instead, we ought to only check
on GET requests. This was causing complications with the car cost
calculator particularly. This commit adds VCR recorded responses from
the CAP API that appear to contain secrets but do not, since the API
requires other secrets to function that are not included in the outputs.