Budget planner retirement (#2528)
* Redirect from legacy tool to new tool syndication

This will allow existing partners to continue syndication of the new
tool in place of the legacy tool, with zero configuration on their side.

* Omit legacy logo when syndicating new tool

The new tool contains its own logo when syndicated so we don't need to
the tools JS to do it as well.

* Remove final canonical redirect

This is no longer needed.

* Redirect from legacy direct download link

Redirects from what was the helper URL for customers to direct download
their completed budgets, to the new tool embed.