Build and deploy using Heroku containers (#2514)
* Remove legacy build cruft

* Use Docker for staging deploys in Heroku

This will build a container using GitHub actions and push to Heroku
staging when a push is made to the `staging` branch.

Resolve tool dependency issues

Bower was incorrectly resolving dependencies and somehow restoring very
old versions of components when building assets. I've traced this down
to an issue in the resolver that is mitigated by using a regular gem
source, rather than github dependencies.

* Install heroku CLI

This isn't part of the runner install now for some reason.

* Deploy to staging or production